% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_config_helper.R
\docType{data}
\name{config_keys}
\alias{config_keys}
\title{Default Configuration Keys}
\format{An object of class \code{list} of length 9.}
\usage{
config_keys
}
\value{
A list with default Redcap configurations as entries.
}
\description{
Default keys for Redcap configuration
}
\details{
This serves as the default set of configurations if no value has been specified for some specific key(s).

Helps in instantiation of RedcapConfig objects.
}
\seealso{
Other Configuration Objects: \code{\link{RedcapConfig}},
  \code{\link{RedcapUpdate}}, \code{\link{load_configs}},
  \code{\link{load_updates}}
}
\keyword{datasets}
