% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_dates_red2r}
\alias{convert_dates_red2r}
\title{Convert REDCap string dates to R dates (code preparation).}
\usage{
convert_dates_red2r(x)
}
\arguments{
\item{x}{a string containing R / REDCap source code.}
}
\value{
Intermediate code with R date syntax that can be evaluated after further modifications into R expression trees
}
\description{
Date conversion from REDCap to R.
}
\details{
In REDCap logic, dates are represented as raw strings.

This is unlike R which requires objects to be of class `Date` for date-like operations to be carried out on the object.

This requires conversion to facilitate date-based operations.
}

