% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_redcap2r}
\alias{convert_redcap2r}
\title{Convert REDCap string dates to valid R code.}
\usage{
convert_redcap2r(x)
}
\arguments{
\item{x}{a string containing REDCap logic.}
}
\value{
a string with containing R code that can be evaluated in its current context
}
\description{
Convert REDCap string dates to valid R code for proper code generation.
}
\details{
REDCap data dictionaries contain logic encapsulated in the branching logic that affects data capture and hence analysis.

These logic has to be transformed into the appropriate analytical tool's DSL for consistent data analysis.

In our case, conversion of REDCap logic to R code.

This is a composite of many smaller internal functions.
}

