% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_space2tab}
\alias{convert_space2tab}
\title{Replace spaces with tabs in code files}
\usage{
convert_space2tab(x)

convert_space2tab(x)
}
\arguments{
\item{x}{a string containing source code}

\item{x}{A string containing R source code}
}
\value{
Properly indented source code

Properly indented R code
}
\description{
Format code by replacing spaces with tabs to allow for block indentation

Format code by replacing spaces with tabs to allow for indentation
}
\details{
For proper code generation, code that is autogenerated must be available for debugging when required.

To allow for this requirement, most of the code has to be indented in a format that the IDE in use would have indented it if it was written manually.

For this, spaces have to be tabbed as required by RStudio, our editor of choice (bias!!).

For proper code generation, code that is autogenerated must be availoable for debugging when required.

To allow for this requirement, most of the code has to be indented in the format the IDE in use would have indented if written manually.

For this, spaces have to be tabbed as required by RStudio, our editor of choice.
}

