% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_dates_red2r}
\alias{convert_dates_red2r}
\title{Convert REDCap string dates to R dates (code preparation).}
\usage{
convert_dates_red2r(x)
}
\arguments{
\item{x}{A string with REDCap logic, either raw or preprocessed.}
}
\value{
Preprocessed R code that can be refined further or valid R syntax that can be parsed into an expression tree for evaluation.
}
\description{
Date conversion from REDCap to R.
}
\details{
In REDCap logic, dates are represented as raw strings.

This is unlike R which requires objects to be of class `Date` for date-like operations to be carried out on the object.

This requires conversion to facilitate date-based operations.
}
\seealso{
Other RedcapToR: \code{\link{convert_missing_red2r}},
  \code{\link{format_branching_logics}},
  \code{\link{logical_xpressions_red2r}},
  \code{\link{xtend_chb_names}}
}
