% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_space2tab}
\alias{convert_space2tab}
\title{Replace leading spaces in code files with tabs}
\usage{
convert_space2tab(x)
}
\arguments{
\item{x}{A string containing source code}
}
\value{
Properly indented source code
}
\description{
Format code by replacing leading spaces with tabs to allow for block indentation
}
\details{
For proper code generation, code that is autogenerated must be available for debugging when required.

To allow for this requirement, most of the code has to be indented in a format that the IDE in use would have indented it if it was written manually.

For this, spaces have to be tabbed as required by RStudio, our editor of choice (bias!!).

This utility function is used to format any custom code for the IDE and is purely for aesthetic purposes.
}
