% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create_updates_info}
\alias{create_updates_info}
\title{Wrapper for creating REDCap update objects}
\usage{
create_updates_info(updates_data)
}
\arguments{
\item{updates_data}{A dataset containing the updates info. For the column specifications, see details.}
}
\value{
A redcap updates class that contains the redcap updates
}
\description{
Cleaner way for instantiating REDCap update objects.

Avoids messing up instantiation in R6 classes which is more complex and error prone.
}
\details{
This function acts as a wrapper for instantiating an object that abstracts the REDCap updates.

It performs the necessary checks for instantiation hence avoids messy objects.

It then calls the new method of the underlying reference class.


The update dataset must contain the following variables:

\describe{
 \item{update}{The update name, eg Nov-14, Feb-13}
 \item{site}{The hospital id of the site, must be integer and is replicated for each update}
 \item{date}{The date of the update for each specific site}
 \item{new_vars}{
 The new variables introduced during the update <must have at least one entry per update>.

 The first non misssing entry will be used.

 The variable names must be separated by semi-colons.
 }
}
}
\seealso{
\code{\link{redcap_update}}, \code{\link{redcap_update_list}}
}

