% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_types.R
\name{DataType_Checks}
\alias{DataType_Checks}
\alias{is_boolean}
\alias{is_date}
\alias{is_int}
\alias{is_number}
\title{Data Type checks for valid string representations of R data types}
\usage{
is_date(x)

is_int(x)

is_number(x)

is_boolean(x)
}
\arguments{
\item{x}{a character vector of length one}

\item{x}{a character vector of length one}

\item{x}{a character vector of length one}

\item{x}{a character vector of length one}
}
\description{
Identify valid data tpes from strings for proper casting








}
\details{
This functions assist in many generic tasks such as configuration type casting and data type checking in error reports.
}

