% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_checks.R
\name{date_can_be_validated}
\alias{data_can_be_validated}
\alias{data_missing}
\alias{date_can_be_validated}
\title{Check whether a date can be used for validation}
\usage{
date_can_be_validated(var)

data_can_be_validated(var)

data_missing(var)
}
\arguments{
\item{var}{The variable to check}
}
\description{
Check whether a date can be used for validation
}
\details{
This performs checks to see if a date can be used for validation in error reporting.

This prevents unnecessary errors and validations resulting from invalid or default data inputs.
}

