% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_branching_logic.R
\name{format_branching_logics}
\alias{format_branching_logics}
\title{Convert Redcap branching logic contained in metadata to R parlance}
\usage{
format_branching_logics(metadata = stop("Provide metadata"),
  data = stop("Provide data"))
}
\arguments{
\item{metadata}{REDCap metadata}

\item{data}{Data associated with metadata}
}
\value{
Transformed metadata with a formatted branching logic in the column name \emph{f.branching_logic}
}
\description{
This is a utility function that converts branching logic to R code.
}
\details{
For code generation, code has to be translated from one DSL or syntax to another.

In this case, conversion from REDCap logic to the appropriate R syntax.

This function helps map logical expressions from REDCap to R
}
\seealso{
Other RedcapToR: \code{\link{convert_dates_red2r}},
  \code{\link{convert_missing_red2r}},
  \code{\link{is_hidden}},
  \code{\link{logical_xpressions_red2r}},
  \code{\link{xtend_chb_names}}
}
