% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_date_conversion_code}
\alias{generate_date_conversion_code}
\title{Autogenerate code for date conversion}
\usage{
generate_date_conversion_code(metadata, dataset_name = "data")
}
\arguments{
\item{metadata}{REDCap metadata}

\item{dataset_name}{Name of the dataset that will be recorded}
}
\value{
Code that can be evaluated to recode dates
}
\description{
This is a utility function that employs code generation to produce r code that recodes data.
}
\details{
Using the redcap metadata, code is generated that converts character dates to R Date variables.
}
\seealso{
Other code_gen: \code{\link{generate_formatting_code}};
  \code{\link{generate_remove_missing_code}};
  \code{\link{generate_remove_outliers_code}}
}

