% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_formatting_code}
\alias{generate_formatting_code}
\title{Autogenerate code for data formatting}
\usage{
generate_formatting_code(metadata, dataset_name = "data")
}
\arguments{
\item{metadata}{REDCap metadata}

\item{dataset_name}{Name of the dataset that will be recorded}
}
\value{
Code that can be evaluated to format data
}
\description{
This is a utility function that employs code generation to produce r code that formats data.
}
\details{
Using the redcap metadata, code is generated that formats data.

It assigns HMisc labels to variables and recodes data to factors
}
\seealso{
Other code_gen: \code{\link{Generate code to eliminate
  missing values}},
  \code{\link{generate_remove_missing_code}};
  \code{\link{Generate code to eliminate out-of-range
  values}}, \code{\link{generate_remove_outliers_code}};
  \code{\link{generate_date_conversion_code}}
}

