% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{generate_remove_missing_code}
\alias{generate_remove_missing_code}
\title{Autogenerate code for removing coded missing values (set to NA) in REDCap data}
\usage{
generate_remove_missing_code(metadata, dataset_name = "data")
}
\arguments{
\item{metadata}{REDCap metadata}

\item{dataset_name}{Name of the dataset that will be recorded}
}
\value{
Code that can be evaluated to clean data
}
\description{
This is a utility function that employs code generation to produce r code for cleaning data.
}
\details{
Using the redcap metadata, code is generated that removes missing data from repo.

This is useful as missing data is coded and has to be recoded as NA.
}
\seealso{
Other code_gen: \code{\link{generate_date_conversion_code}};
  \code{\link{generate_formatting_code}};
  \code{\link{generate_remove_outliers_code}}
}

