% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{generate_data_validation_code}
\alias{generate_data_validation_code}
\title{Autogenerate code for data validation}
\usage{
generate_data_validation_code(metadata, date_var, hosp_var, surrogate_id_var,
  custom_code = NA, updates = NULL, updates_envir_depth = 1,
  negative_char = "_")
}
\arguments{
\item{metadata}{REDCap metadata with a formatted branching logic variable \emph{f.branching_logic}. See \code{\link{ExpandBranchingLogic}}}

\item{date_var}{Name of variable that captures the date of entry}

\item{hosp_var}{Name of variable that holds the hospital code}

\item{surrogate_id_var}{Name of variable that holds a surrogate identifier that is easier to reference}

\item{custom_code}{Any code that is appended for custom plugin of special validation checks.}

\item{updates}{Name of a list of RedcapUpdate(s) to be used for plugging functionality that abstracts the introduction of new variables during the projects lifecycle. See \code{\link{RedcapUpdate}}}

\item{updates_envir_depth}{Integer of what parent frame contains updates. Default is immediate parent of calling environment (1) ie one level deep.}

\item{negative_char}{checkbox expansion character for a negative checkbox level}
}
\value{
A function that can be evaluated to validate data entry workflow for a single record.
}
\description{
This is a utility function that employs code generation to produce R code that validates data entry workflow for errors of omission and commision.
}
\details{
Using the redcap metadata, code is generated that validates data entry during the data capture process.

This code is then evaluated into a function that is then iterated through the records to check for errors during data capture.

The result is a dataset containing the resultant errors.
}
\seealso{
Other Code Generators: \code{\link{generate_date_conversion_code}},
  \code{\link{generate_formatting_code}},
  \code{\link{generate_remove_missing_code}},
  \code{\link{generate_remove_outliers_code}}
}
