% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{generate_remove_outliers_code}
\alias{generate_remove_outliers_code}
\title{Autogenerate code for removing out of range values (set to NA) in REDCap data}
\usage{
generate_remove_outliers_code(metadata, dataset_name = "data")
}
\arguments{
\item{metadata}{REDCap metadata}

\item{dataset_name}{Name of the dataset that will be recorded}
}
\value{
Code that can be evaluated to clean data
}
\description{
This is a utility function that employs code generation to produce r code for cleaning data.
}
\details{
Using the redcap metadata, code is generated that removes out-of-range data from repo.

This is necessary to avoid over or undeestimation during data analysis which often distorts the results.
}
\seealso{
Other Code Generators: \code{\link{generate_data_validation_code}},
  \code{\link{generate_date_conversion_code}},
  \code{\link{generate_formatting_code}},
  \code{\link{generate_remove_missing_code}}
}
