% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Get REDCap data}
\alias{Get REDCap data}
\alias{get_redcap_data}
\title{Pull data from REDCap}
\usage{
get_redcap_data(api, token, content = "record", local = TRUE,
  forms = NULL, fields = NULL, ids_to_pull = NULL)
}
\arguments{
\item{api}{the REDCap instance's api location. Just append /api/ to the instance's url.}

\item{token}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{content}{What to pull. Currently only "record" and "metadata" are supported.}

\item{local}{Whether the REDCap instance is local.}

\item{forms}{a character vector of the list of form data to pull.}

\item{fields}{a character vector of the specific fields to pull.}

\item{ids_to_pull}{a character vector of the specific record itentifiers to pull.}

\item{chunksize}{The size of the chunks to be pulled at a time.}
}
\value{
A data frame with the REDCap repo's data.
}
\description{
Download data from a REDCap repository using the REDCap api in bulk.
}
\details{
A simulation of the posting of a form to the api is done and a connection is obtained.

The connection is then used to read the data into REDCap.

This is done in bulk. In the case  of network latency or big data \code{\link{get_chunked_redcap_data}} would be a good alternative.
}
\seealso{
\code{\link{get_chunked_redcap_data}}, \code{\link{redcap}}
}

