% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Get variables in dataset}
\alias{Get variables in dataset}
\alias{get_vars_in_data}
\title{Get the names of variables in dataset}
\usage{
get_vars_in_data(metadata)
}
\arguments{
\item{metadata}{REDCap metadata}
}
\value{
a character vector of the names of the dataset
}
\description{
Utility function that gets the name(s) of the variables in a redcap project.
}
\details{
Using the redcap metadata, code is generated that extracts the names of the dataset.

Checkboxes are also munged to reflect what is in the repository.
}

