% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_r_types_in_data}
\alias{get_r_types_in_data}
\title{Get the r data types of variables in dataset}
\usage{
get_r_types_in_data(metadata)
}
\arguments{
\item{metadata}{REDCap metadata}
}
\value{
The data types (r) of the variables in the dataset
}
\description{
Utility function that gets the data type(s) of the variables in a redcap project
}
\details{
Using the redcap metadata, code is generated that extracts the data types of the variables in the dataset.
}

