% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_redcap_data}
\alias{get_redcap_data}
\title{Bulk download of REDCap data}
\usage{
get_redcap_data(api, token, content = "record", local = TRUE,
  forms = NULL, fields = NULL, ids_to_pull = NULL)
}
\arguments{
\item{api}{The REDCap instance's api location. Just append /api/ to the instance's url.}

\item{token}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{content}{What to pull. Currently only `record` and `metadata` are supported.}

\item{local}{Whether the REDCap instance is local.}

\item{forms}{A character vector of the list of forms to pull data from.}

\item{fields}{A character vector of the specific fields to pull data from.}

\item{ids_to_pull}{A character vector of the specific record itentifiers to pull data from.}
}
\value{
A data frame with the REDCap repo's data.
}
\description{
Download data from a REDCap repository using the REDCap api in bulk.
}
\details{
A simulation of the posting of a form to the api is done and a connection is obtained.

This connection is then used to read the data from REDCap into R.

This is done in bulk. In the case  of network latency or big data \code{\link{get_chunked_redcap_data}} would be a better alternative for more responsiveness.
}
\seealso{
\code{\link{redcap}}

Other Data Input: \code{\link{get_chunked_redcap_data}}
}

