% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_status}
\alias{get_status}
\title{Get Cache status}
\usage{
get_status(cache_objects, pretty = FALSE)
}
\arguments{
\item{cache_objects}{Redcap object's cache}

\item{pretty}{Whether to format output for display}
}
\value{
Code that can be evaluated to format data
}
\description{
From the cache in the redcap object, identify the major events that have happened.
}
\details{
This function helps format the cache entries so as to provide a meaningful decmdion of the events that happened during the object' lifecycle.

This also helps in formatting output in the show command
}

