% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_status}
\alias{get_status}
\title{Get Cache status}
\usage{
get_status(cache_objects, pretty = FALSE)
}
\arguments{
\item{cache_objects}{Redcap object's cache}

\item{pretty}{Whether to format output for display}
}
\value{
Code that can be evaluated to format cache status
}
\description{
From the cache in the redcap object, identify the major events that have ocurred during the lifetime of a Redcap object
}
\details{
This function helps format the cache entries so as to provide a meaningful description of the events that happened during the object' lifecycle.

This also helps in formatting output in the show command.
}
