% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_chunked_redcap_data}
\alias{get_chunked_redcap_data}
\title{Download REDCap data in chunks}
\usage{
get_chunked_redcap_data(api, token, local = TRUE, chunksize = 100,
  forms = NULL, fields = NULL, ids_to_pull = NULL,
  dataset_name = "records", metadataset_name = "meta")
}
\arguments{
\item{api}{the REDCap instance's api location. Just append /api/ to the instance's url.}

\item{token}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{local}{Whether the REDCap instance is local.}

\item{chunksize}{The size of the chunks to be pulled at a time.}

\item{forms}{A character vector of the list of forms to pull data from.}

\item{fields}{A character vector of the specific fields to pull data from.}

\item{ids_to_pull}{A character vector of the specific record itentifiers to pull from.}

\item{dataset_name}{Name of the resultant dataset}

\item{metadataset_name}{Name of the resultant metadata}
}
\description{
Download data in chunks from a REDCap repository using the REDCap api.

Data and metadata are stored in the calling environment.
}
\details{
The record identifiers are chunked and then data is is pulled for those specific records.

The result is then combined into one data table.

This might be handy in the case of network latency or when the data is large.

This is just a convenient wrapper round the \code{\link{get_redcap_data}} function.

In cases of strong bandwidth or small data sizes, just use \code{\link{get_redcap_data}}.
}
\seealso{
\code{\link{get_chunks}}, \code{\link{Redcap}}

Other Data Input: \code{\link{get_redcap_data}}
}
