% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_chunks}
\alias{get_chunks}
\title{Get chunks of a specific size}
\usage{
get_chunks(x, chunksize)
}
\arguments{
\item{x}{Input vector}

\item{chunksize}{The size of the chunks}
}
\value{
A list of objects of the same data type as input, each of size equal to the chunk size.

The last chunk may be smaller if the input's length is not a multiple of the LCM of the input's length and the chunks size.
}
\description{
From a vector obtain a list of chunks of a specific size
}
\details{
This is a utility functiion that facilitates chunked operations.

A common scenario is to chunk the indices of a specific object and then perform operations on the chunks separately and then later merge the output.

This can be useful in situations where multiple operations of smaller scope is preferrable to bulk processing.
}
\seealso{
\code{\link{get_chunked_redcap_data}}
}
