% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_config_helper.R
\name{get_config_type_from_variable}
\alias{get_config_type_from_variable}
\title{Get configuration type from variable}
\usage{
get_config_type_from_variable(configs_list)
}
\arguments{
\item{x}{A list of configurations with the names set to the corresponding key}
}
\value{
A string of the special config data type
}
\description{
From an R object set at configuration, define the required special type for configuration
}
\details{
This is a utility function that helps map R types to their corresponding config type as expected for this package.

This allievates the use of csv files for configuration as options can be specified as anonymous function arguments (kwargs)
}
\seealso{
\code{\link{redcap_project}}
}
