% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_redcap_version}
\alias{get_redcap_version}
\title{Get REDCap version}
\usage{
get_redcap_version(url = "http://localhost/redcap")
}
\arguments{
\item{url}{The url of the REDCap instance}
}
\value{
A list containing the version info

This includes the major, minor and release number version information for the REDCap instance
}
\description{
Get the version of REDCap associated with a specific url
}
\details{
This function gets the REDCap version number by web scraping the specific instance's index page.

The footer is extracted which contains the version number.

This is then wranged into the final version information.
}
\seealso{
\code{\link{redcap_project}}
}
