% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{get_vars_in_data}
\alias{get_vars_in_data}
\title{Get the names of variables in a dataset based on REDCap metadata}
\usage{
get_vars_in_data(metadata)
}
\arguments{
\item{metadata}{REDCap metadata}
}
\value{
a character vector of the variable names in the dataset
}
\description{
Utility function that gets the name(s) of the variables in a redcap project.
}
\details{
Using the redcap metadata, code is generated that extracts the names of the variables in the dataset.

Checkboxes are also munged to reflect what is in the repository.

Useful for metaprogramming.
}
