% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Is Metadata Valid}
\alias{Is Metadata Valid}
\alias{is_valid_metadata}
\title{Check whether metadata is valid}
\usage{
is_valid_metadata(metadata)
}
\arguments{
\item{metadata}{REDCap metadata}
}
\value{
TRUE if valid else FALSE
}
\description{
This performs a check to validate the REDCap metadata provided.
}
\details{
This makes sure that the REDCap metadata conforms to the expectations of this project.

This helps avoid breaking of code in instances where metadata is required as an iput
}

