% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_configs}
\alias{load_configs}
\title{Wrapper for creating REDCap configuration objects}
\usage{
load_configs(config_data = NULL, custom_code = NA, exclusion_pattern = NA)
}
\arguments{
\item{updates_data}{A dataset containing the updates info. For the column specifications, see details.}
}
\value{
A list of redcap update class objects.
}
\description{
Cleaner way for instantiating REDCap configuration objects.

It uses an R data frame and character vectors to instantiate REDCap configuration objects.

This avoids messing up instantiation in R6 classes which is more complex and error prone.
}
\details{
This function acts as a wrapper for instantiating an object that abstracts the REDCap updates.

It performs the necessary checks for instantiation hence avoids messy objects.

It then calls the new method of the underlying reference class.

The update dataset must contain the following variables:

\describe{
 \item{update}{The update name of format mon-yy, eg Nov-14, Feb-13}
 \item{site}{The hospital id of the site, must be integer and must be replicated for each update}
 \item{date}{The date of the update for each specific site}
 \item{new_vars}{
 The new variables introduced during the update <must have at least one entry per update>.

 The first non misssing entry will be used.

 The variable names must be separated by semi-colons.
 }
}
}
\seealso{
\code{\link{redcap_update}}
}

