% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{logical_xpressions_red2r}
\alias{logical_xpressions_red2r}
\title{Reshape REDCap logical expresssions to R logical code.}
\usage{
logical_xpressions_red2r(x)
}
\arguments{
\item{x}{string with REDCap logic}
}
\value{
R code that can be evaluated after further modifications
}
\description{
Convert REDCap logical expressions to valid R syntax representations.
}
\details{
For code generation, code has to be translated from one DSL or syntax to another.

In this case, conversion from REDCap logic to the appropriate R syntax.

This function helps map logical expressions from REDCap to R
}

