% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_helper.R
\name{open_using_default_app}
\alias{open_using_default_app}
\title{Open file using OS default application}
\usage{
open_using_default_app(file_path)
}
\arguments{
\item{file_path}{Path to target file}
}
\description{
Pop a file open using the default application set at the operating system level to handle files of the type specified.
}
\details{
This allows for results to be viewed interactively during analysis and programming.

This is only applicable in GUI settings and should only be used in windowed environmnents.
}
