% Generated by roxygen2 (4.0.2): do not edit by hand
\name{prepare_metadata_for_code_generation}
\alias{prepare_metadata_for_code_generation}
\title{Prepare metadata for code generation}
\usage{
prepare_metadata_for_code_generation(metadata)
}
\arguments{
\item{metadata}{Redcap metadata}
}
\value{
transformed metadata
}
\description{
Take metadata and make sure it conforms to the project's meta programming.
}
\details{
This is a utility function that aids in preparing the metadata for cmd generation.
It converts the iput to a data table and assigns a key to it.
}

