% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{prepare_metadata_for_code_generation}
\alias{prepare_metadata_for_code_generation}
\title{Prepare metadata for code generation}
\usage{
prepare_metadata_for_code_generation(metadata)
}
\arguments{
\item{metadata}{REDCap metadata}
}
\value{
transformed metadata
}
\description{
Take metadata and make sure it conforms to the project's metaprogramming expected format.
}
\details{
This is a utility function that aids in preparing the metadata for cmd generation.

It converts the input to a data table and assigns a key to it and also removes unnecessary fields.
}
