% Generated by roxygen2 (4.0.2): do not edit by hand
\name{redcap}
\alias{redcap}
\title{Wrapper for creating REDCap objects}
\usage{
redcap(token, api_url = "http://localhost/redcap/api/", local = TRUE,
  exclusion_pattern = NA_character_, custom_error_file = NA_character_)
}
\arguments{
\item{token}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{api_url}{REDCap project's api url. Just append /api/ to the project's url.}

\item{local}{Whether REDCap instance is local.}

\item{exclusion_pattern}{regex for variables to exclude from autogeneration of code}

\item{custom_code}{Custom error report code}
}
\value{
A redcap class instance that can be used to interact with the data repository
}
\description{
Cleaner way for instantiating REDCap objects.

Avoids messing up instantiation in R6 classes which is more complex and error prone.
}
\details{
This function acts as a wrapper for instantiating an object that abstracts the REDCap API.

It performs the necessary checks for instantiation hence avoids messy objects.

It then calls the new method of the underlying reference class.
}
\section{Info}{

Use redcap function to instantiate this class. This avoids many pitfalls during the lifetime of this object.
}
\examples{
\dontrun{cin = redcap("<some-token>", api_url = "http://<some-dns>/redcap/api/", local = F)}

\dontrun{cin}
}
\seealso{
\code{\link{redcap_class}}
}

