% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{redcap_class}
\alias{cin_project}
\alias{redcap_class}
\alias{redcap_project}
\alias{redcap_wrapper}
\title{Redcap Wrapper Class}
\value{
A redcap class with an array of functionality
}
\description{
This class is the entry point for interacting with the CIN data which is stored in a REDCap data repository
}
\details{
This object facilitates R's interface with the REDCap repository.

Data can be pulled either in bulk or chunked.

Currently only metadata and records can be pulled.

It also provides other data allied tasks such as data cleaning, error reporting and formatting.

It uses a cache system and logs major events

Show the object to view status
}
\section{Fields}{

\describe{
\item{\code{api}}{The url to the project's web based api. Just append /api/ to the project's url.}

\item{\code{token}}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{\code{local}}{Whether REDCap instance is local}

\item{\code{exclusion_pattern}}{regex for variables to exclude from autogeneration of code}

\item{\code{custom_code}}{Custom error report code}
}}
\section{Methods}{

\describe{
\item{\code{clean_meta(is_error = FALSE)}}{Clean meta data for autogeneration of cleaning code}

\item{\code{clean_records()}}{Clean rceords removing out of range and empty values}

\item{\code{format_records()}}{format records to add Hmisc labels and create factors}

\item{\code{get_clean_data()}}{Get saved records from cache}

\item{\code{get_error_report(save_to = NULL, pop = TRUE)}}{Get error report}

\item{\code{get_formatted_data()}}{Get saved records from cache}

\item{\code{get_metadata()}}{Get saved metadata from cache}

\item{\code{get_raw_data()}}{Get saved records from cache}

\item{\code{load_data(chunked = FALSE, chunksize = NULL, forms = NULL,
  fields = NULL, ids_to_pull = NULL)}}{Stream data from REDCap if not in cache}

\item{\code{load_metadata()}}{Stream metadata only from REDCap if not in cache}

\item{\code{log(message, level = 0, function_name = "")}}{Log events <internal>}

\item{\code{report_errors()}}{Create error report}
}}
\section{Info}{

Use redcap function to instantiate this class. This avoids many pitfalls during the lifetime of this object.
}
\seealso{
\code{\link{redcap}}
}
\concept{
cin
}

