% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{redcap_config}
\alias{redcap_config}
\title{Redcap Configuration}
\value{
A RedcapConfig class
}
\description{
This class holds redcap configurations that control the interaction with the data repository.
}
\details{
This configurations affect various processes encapsulated by the REDCap class from data input to error reporting.

This allows the configurations to be loosely coupled from the REDCap interaction for both flexibility and maintainability.
}
\section{Fields}{

\describe{
\item{\code{updates}}{a list of update objects.}

\item{\code{configs}}{an internal environment for holding configurations.}
}}
\section{Methods}{

\describe{
\item{\code{configs()}}{Displays the configurations for the REDCap session/object}

\item{\code{get_update_date(var_name, hospital_id)}}{Get the date a specific variable was updated for a specific site}

\item{\code{is_valid()}}{Checks the validity of the object}
}}
\seealso{
Other config_objects: \code{\link{redcap_update}}
}
\concept{
configuration
}

