% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{redcap_update_list}
\alias{redcap_update_list}
\title{Redcap Updates Collection}
\value{
A redcap update collection class
}
\description{
This class holds a collection of update info objects that control the error reporting process.
}
\details{
It holds the collection of objects that control the error reporing code generation.

It also has collection-based operations.
}
\section{Fields}{

\describe{
\item{\code{updates}}{a list of update objects.}
}}
\section{Methods}{

\describe{
\item{\code{get_update_date(var_name, hospital_id)}}{Get the date a specific variable was updated for a specific site}

\item{\code{is_valid()}}{Checks the validity of the object}
}}
\seealso{
Other update objects: \code{\link{redcap_update}}
}
\concept{
update redcap
}

