% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_config.R
\docType{class}
\name{RedcapConfig}
\alias{RedcapConfig}
\title{Redcap Configuration Wrapper Class}
\value{
A Redcap configuration class
}
\description{
This class holds redcap configurations that control the mode of interaction with the data repository.
}
\details{
This configurations affect various processes encapsulated by the REDCap class from data input to error reporting.

This allows the configurations to be loosely coupled from the REDCap interaction for both flexibility and maintainability.
}
\section{Fields}{

\describe{
\item{\code{updates}}{a list of RedcapUpdate object(s).}

\item{\code{configs}}{an internal environment for holding configurations.}
}}
\section{Methods}{

\describe{
\item{\code{get_update_date(var_name, hospital_id)}}{Get the date a specific variable was updated for a specific site}

\item{\code{is_valid()}}{Checks the validity of the object}

\item{\code{list_configs()}}{Display the configurations for the REDCap session/object}
}}
\seealso{
Other Configuration Objects: \code{\link{RedcapUpdate}};
  \code{\link{config_keys}}; \code{\link{load_configs}};
  \code{\link{load_updates}}
}
\concept{
configuration cin redcap
}

