% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_project}
\alias{redcap_project}
\alias{cin_project}
\title{Wrapper for creating REDCap objects}
\usage{
redcap_project(..., configs_location, custom_code_location = NA,
  updates_location = NA, exclusion_pattern = NA_character_)
}
\arguments{
\item{...}{Individual configuration settings as expected in the configs file.

If \code{configs_location} is specified, these are discarded.

See details...}

\item{configs_location}{Location of the configs file (csv). See details...}

\item{custom_code_location}{Location of any custom code for error reporting (.R | .txt). See details...}

\item{updates_location}{Location of a file containing any updates to redcap metadata. (csv) See details...}

\item{exclusion_pattern}{A character vector of regex patterns for variables to exclude from autogeneration of error reporting code.}
}
\value{
A redcap class instance that can be used to interact with the data repository
}
\description{
Cleaner way for instantiating REDCap objects.

Avoids messing up instantiation in R6 classes which is more complex and error prone.
}
\details{
This function acts as a wrapper for instantiating an object that abstracts the REDCap API.

It performs the necessary checks for instantiation hence avoids messy objects.

It then calls the new method of the underlying reference class.

The configs and update files must be csv files. See \code{\link{load_configs}}, \code{\link{load_updates}} for details.

The custom code must hold valid R code.

The exclusion patterns must hold valid R regex expressions. Can be full variable names or a set of patterns.
}
\seealso{
\code{\link{Redcap}}
}
