% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_config.R
\docType{class}
\name{RedcapUpdate}
\alias{RedcapUpdate}
\title{Redcap Update Wrapper Class}
\value{
A redcap update class
}
\description{
This class holds the update information that affects the error reporting process.
}
\details{
It holds the site information and the date of updates hence it is possible to do site-based operations.

The dates can be accessed to allow for conditional generation of the error reports.
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the update.}

\item{\code{site_info}}{A dataset of site information. It has variables `site` and `date` matching the site identifier and date of the update for each site.}

\item{\code{new_vars}}{New variables introduced during the update.}
}}
\section{Methods}{

\describe{
\item{\code{get_update_date(var_name, hospital_id)}}{Get the date a specific variable was updated for a specific site}

\item{\code{is_valid()}}{Check validity of object}
}}
\seealso{
Other Configuration Objects: \code{\link{RedcapConfig}};
  \code{\link{load_configs}}; \code{\link{load_updates}}
}
\concept{
updates configuration cin redcap
}

