% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tab_utils}
\alias{add_tab}
\alias{get_tab}
\alias{remove_tab}
\alias{reset_tab}
\alias{tab_utils}
\title{Manage Tabbing in script creation}
\usage{
add_tab(n = 1)

remove_tab(n = 1)

reset_tab()

get_tab()
}
\description{
Indent code for easier reading and debugging.
}
\details{
These help in code generation.

Proper use of these functions result in well indented scripts during metaprogrammiung.
}

