% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{TabUtils}
\alias{TabUtils}
\alias{add_tab}
\alias{remove_tab}
\alias{reset_tab}
\alias{get_tab}
\title{Manage tabs when autogenerating scripts}
\usage{
add_tab(n = 1)

remove_tab(n = 1)

reset_tab()

get_tab()
}
\arguments{
\item{n}{Number of tabs to add / remove}
}
\description{
Indent code for easier reading and debugging especially in IDEs.
}
\details{
These help in code generation when metaprogramming.

Proper use of these functions result in well indented scripts.
}
