% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fields_without_ranges.R
\name{validate_data_in_branching_logic}
\alias{validate_data_in_branching_logic}
\title{Generate custom codes and evaluate based on branching logics}
\usage{
validate_data_in_branching_logic(rec, metadataName = "metadata.formatted",
  dataName = "data.raw",
  ipno_var = stop("Provide variable name for IP Number"),
  dateOfEntry_var = stop("Provide variable name for date of entry"),
  recordID_var = stop("Provide variable name for record ID"),
  hospitalID_var = stop("Provide variable name for hospitals"),
  individual.vars = stop("Provide a list of variables to be validated individually"),
  n.groups = NULL, group.names = NULL, validateTreatmentDates = F,
  treatmentFormName = NULL, dischargeDateVar = NULL,
  admissionDateVar = NULL)
}
\arguments{
\item{rec}{Single record in the data}

\item{metadataName}{Metadata name}

\item{dataName}{Data name}

\item{ipno_var}{Name of variable for patient IP Number}

\item{dateOfEntry_var}{Name of variable that captures the date of entry}

\item{recordID_var}{Name of variable that uniquely identifies records}

\item{hospitalID_var}{Name of variable that holds the hospital code}

\item{individual.vars}{A \emph{character} name of the object containing variables to be validated individually}

\item{n.groups}{Numebr of groups provided in \strong{group.names}}

\item{group.names}{A vector of \emph{character} names of the objects containing variables to be validated as a group}

\item{validateTreatmentDates}{Logical. If FALSE, treatment dates will not be validated}

\item{treatmentFormName}{Name of the treatment section in the metadata. This is required if \strong{validateTreatmentDates} is TRUE}

\item{dischargeDateVar}{Name of variable that holds the date of discharge. This is required if \strong{validateTreatmentDates} is TRUE}

\item{dischargeDateVar}{Name of variable that holds the date of admission. This is required if \strong{validateTreatmentDates} is TRUE}
}
\value{
A data frame with details of the errors.
}
\description{
This is a utility function that uses branching logic to validate data entry workflow for errors of omission.
}
\details{
Using the redcap metadata's newly created variable \emph{f.branching_logic}, code is generated that validates data entry during the data capture process.
}
