% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_config_helper.R
\docType{data}
\name{config_keys}
\alias{config_keys}
\title{Default Configuration Keys}
\format{\preformatted{List of 9
 $ api_url         : chr "http://localhost/redcap/api/"
 $ token           : chr ""
 $ local           : logi TRUE
 $ chunked         : logi FALSE
 $ chunksize       : int NA
 $ hosp_var        : chr "hosp_id"
 $ date_var        : chr "date_today"
 $ report_location : chr "C:\\\\Users\\\\bmakone\\\\AppData\\\\Local\\\\Temp\\\\RtmpoRqSzf\\\\Error_Report1188b44237a.csv"
 $ hosp_to_validate: int NA
}}
\usage{
config_keys
}
\description{
Default configuration keys
}
\details{
This serves as the default set of configurations if no value has been specified for som specific key.

Helps in instantiation of RedcapConfig objects
}
\seealso{
\code{\link{RedcapConfig}}
}
\keyword{datasets}

