% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_dates_red2r}
\alias{convert_dates_red2r}
\title{Convert REDCap string dates to R dates (code preparation).}
\usage{
convert_dates_red2r(x)
}
\arguments{
\item{x}{a string containing R / REDCap source code.}
}
\value{
a string with R/ REDCap source code that has raw string dates replaced with formatted dates.
}
\description{
Date conversion from REDCap to R.
}
\details{
In REDCap logic, dates are represented as raw strings.

This is unlike R which requires objects to be of class `Date` for date-like operations to be carried out on the object.

This is why conversion to date-operations compatible code is necessary.
}

