% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convert_missing_red2r}
\alias{convert_missing_red2r}
\title{Reshape REDCap missingness logic to appropriate R code}
\usage{
convert_missing_red2r(x)
}
\arguments{
\item{x}{string with REDCap logic}
}
\value{
R code that can be evaluated after further modifications
}
\description{
Convert REDCap missing code to corresponding R syntax.
}
\details{
For code generation, code has to be translated from one DSL or syntax to another.

In this case, conversion from REDCap logic to the appropriate R syntax.

This function helps map logical expressions from REDCap to R
}

