% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{convert_redcap2r}
\alias{convert_redcap2r}
\title{Convert REDCap logic to R code.}
\usage{
convert_redcap2r(x)
}
\arguments{
\item{x}{a string containing REDCap logic.}
}
\value{
a string with containing R code that can be evaluated in its current context
}
\description{
Wrapper for converting REDCap branching logic to valid R code (metaprogramming purposes).
}
\details{
REDCap data dictionaries contain logic encapsulated in the branching logic that affects data capture and hence analysis.

These logic has to be transformed into the appropriate analytical tool's DSL for consistent data analysis.

In our case, conversion of REDCap logic to R code.

This is a composite of many smaller internal functions.
}

