% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_types.R
\name{DataTypeChecks}
\alias{DataTypeChecks}
\alias{is_date}
\alias{is_int}
\alias{is_number}
\alias{is_boolean}
\title{Data Type checks for valid string representations of R data types}
\usage{
is_date(x)

is_int(x)

is_number(x)

is_boolean(x)
}
\arguments{
\item{x}{a character vector of unit length}
}
\description{
Identify valid R data types from strings for proper type casting during data analysis and munging
}
\details{
This functions assist in many generic tasks such as configuration type casting, branching logic syntax conversion and data type checking during error reporting.

They are simply regular expressions that are evaluated to boolean outputs depending on whether the predicate meets the requirement.
}
