% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap_helper.R
\name{generate_formatting_code}
\alias{generate_formatting_code}
\title{Autogenerate code for data formatting (variable and data labelling)}
\usage{
generate_formatting_code(metadata, dataset_name = "data")
}
\arguments{
\item{metadata}{REDCap metadata}

\item{dataset_name}{Name of the dataset that will be recorded in place.}
}
\value{
Code that can be evaluated to format data.
}
\description{
This is a utility function that employs code generation to produce R code that formats data.
}
\details{
Using the redcap metadata, code is generated that formats data.

It assigns HMisc labels to variables and recodes data to factors.

This makes it easier to perform traditional statistical analysis which often expects coded categorical variables as input.
}
\seealso{
Other Code Generators: \code{\link{generate_date_conversion_code}};
  \code{\link{generate_error_report_code}};
  \code{\link{generate_remove_missing_code}};
  \code{\link{generate_remove_outliers_code}}
}

