% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_chunked_redcap_data}
\alias{get_chunked_redcap_data}
\title{Perform a chunked REDCap data dowmload}
\usage{
get_chunked_redcap_data(api, token, local = TRUE, chunksize = 100,
  forms = NULL, fields = NULL, ids_to_pull = NULL,
  dataset_name = "records", metadataset_name = "meta")
}
\arguments{
\item{api}{the REDCap instance's api location. Just append /api/ to the instance's url.}

\item{token}{The secret token for the project. Check the API page in REDCap. Must have api rights to access this.}

\item{local}{Whether the REDCap instance is local.}

\item{chunksize}{The size of the chunks to be pulled at a time.}

\item{forms}{a character vector of the list of form data to pull.}

\item{fields}{a character vector of the specific fields to pull.}

\item{ids_to_pull}{a character vector of the specific record itentifiers to pull.}

\item{dataset_name}{name of the resultant dataset}

\item{metdataset_name}{name of the resultant metadata}
}
\description{
Download data in chunks from a REDCap repository using the REDCap api.

If not specified, data and metadata are stored in the current environment.
}
\details{
The record identifiers are chunked and then data is is pulled for those specific records.

The result is then combined into one data table.

This might be handy in the case of network latency or when the data is large.

This is just a convenient wrapper round the \code{\link{get_redcap_data}} function.
}
\seealso{
\code{\link{get_chunks}}

Other get_data_red: \code{\link{get_redcap_data}}
}

