% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\docType{class}
\name{Redcap}
\alias{Redcap}
\alias{redcap_wrapper}
\alias{redcap_class}
\title{Redcap Wrapper Class}
\value{
A redcap class with an array of data-allied functionality.
}
\description{
This class is the entry point for interacting with data stored in a REDCap data repository
}
\details{
This object facilitates R's interface with the REDCap repository.

Data can be pulled either in bulk or chunked.

Currently only metadata and records can be pulled.

It also provides other data allied tasks such as data cleaning, error reporting and formatting.

It uses a cache system and logs major events.

Show the object to view status.
}
\section{Fields}{

\describe{
\item{\code{opts}}{RedcapConfig object that controls the mode of interaction with the data repoistory during this object's lifecycle.}
}}

\section{Methods}{

\describe{
\item{\code{format_records()}}{Format records to add Hmisc labels and create factors}

\item{\code{fully_clean_records()}}{Clean records removing out of range values}

\item{\code{get_clean_metadata()}}{Clean meta data for autogeneration of error report code}

\item{\code{get_data_validation_report(pop = FALSE)}}{Get data validation report}

\item{\code{get_formatted_data()}}{Get formatted records from memory}

\item{\code{get_fully_cleaned_data()}}{Get fully cleaned records from memory}

\item{\code{get_metadata()}}{Get raw metadata from memory}

\item{\code{get_partially_cleaned_data()}}{Get partially cleaned records from memory}

\item{\code{get_raw_data()}}{Get raw records from memory.
If there is not data in memory, an error is raised.
Use load_data() to refresh the cache.}

\item{\code{load_data()}}{Stream data from REDCap.
Clears the cache and loads data.
< NOTE: Do this only for initial loading or when you are sure there are changes in the data repo >.}

\item{\code{load_metadata()}}{Stream metadata only from REDCap}

\item{\code{log(message, level = 0, function_name = "")}}{Log events <internal use>}

\item{\code{partially_clean_records()}}{Clean records removing coded missing values}

\item{\code{validate_data()}}{Create data validation dataset}
}}

\seealso{
\code{\link{redcap_project}}
}
\concept{
redcap_data
}
