% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/redcap.R
\docType{class}
\name{Redcap}
\alias{Redcap}
\alias{redcap_class}
\alias{redcap_wrapper}
\title{Redcap Wrapper Class}
\value{
A redcap class with an array of data-allied functionality
}
\description{
This class is the entry point for interacting with the CIN data which is stored in a REDCap data repository
}
\details{
This object facilitates R's interface with the REDCap repository.

Data can be pulled either in bulk or chunked.

Currently only metadata and records can be pulled.

It also provides other data allied tasks such as data cleaning, error reporting and formatting.

It uses a cache system and logs major events.

Show the object to view status.
}
\section{Fields}{

\describe{
\item{\code{opts}}{RedcapConfig object that controls the mode of interaction with the data repoistory during this objects lifecycle}
}}
\section{Methods}{

\describe{
\item{\code{clean_records()}}{Clean records removing out of range and coded missing values}

\item{\code{format_records()}}{Format records to add Hmisc labels and create factors}

\item{\code{get_clean_data()}}{Get cleaned records from memory}

\item{\code{get_clean_metadata(is_error = FALSE)}}{Clean meta data for autogeneration of error report code}

\item{\code{get_error_report(save_to = NULL, pop = TRUE)}}{Get error report}

\item{\code{get_formatted_data()}}{Get formatted records from memory}

\item{\code{get_metadata()}}{Get raw metadata from memory}

\item{\code{get_raw_data()}}{Get raw records from memory.
If there is not data in memory, an error is raised.
Use load_data() to refresh the cache.}

\item{\code{load_data()}}{Stream data from REDCap.
Clears the cache and loads data.
< NOTE: Do this only for initial loading or when you are sure there are changes in the data repo.}

\item{\code{load_metadata()}}{Stream metadata only from REDCap}

\item{\code{log(message, level = 0, function_name = "")}}{Log events <internal use>}

\item{\code{report_errors()}}{Create error report}
}}
\section{Info}{

Use `redcap_project` function to instantiate this class. This avoids many pitfalls during the lifetime of this object.
}
\seealso{
\code{\link{redcap_project}}
}
\concept{
cin
}

