% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{redcap_update}
\alias{redcap_update}
\title{Redcap Update Object}
\value{
A redcap update class
}
\description{
This class holds the update information that affects the error reporting process.
}
\details{
It holds the site information and the date of updates hence it is possible to do site-based operations.

The dates can be accessed to allow for conditional generation of the error reports.
}
\section{Fields}{

\describe{
\item{\code{name}}{Name of the update.}

\item{\code{site_info}}{A dataset of site information. It has variables `site` and `date` matching the date of the update for each site.}

\item{\code{new_vars}}{New variables added during the update.}
}}
\section{Methods}{

\describe{
\item{\code{get_update_date(var_name, hospital_id)}}{Get the date a specific variable was updated for a specific site}
}}
\seealso{
Other config_objects: \code{\link{redcap_config}}
}
\concept{
updates configuration
}

