% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{remove_html_tags}
\alias{remove_html_tags}
\title{Remove html tags from string}
\usage{
remove_html_tags(x)
}
\arguments{
\item{x}{string to refine}
}
\value{
A html free string
}
\description{
Removes html tags from string which can later be used as html-free text eg labels.
}
\details{
This is a utility function for refining text.

This is useful in generic tasks such as autogeneration of code.
}

