% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{remove_html_tags}
\alias{remove_html_tags}
\title{Remove html tags from string}
\usage{
remove_html_tags(x)
}
\arguments{
\item{x}{A string possibly with embedded html tags.}
}
\value{
A string free of html tags and elements.
}
\description{
Removes html tags from string which can later be used as html-free text e.g. in labels.
}
\details{
This is a utility function for refining text.

This is useful in generic tasks such as autogeneration of code.
}
