% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/script.R
\name{Tab_Utils}
\alias{Tab_Utils}
\alias{add_tab}
\alias{get_tab}
\alias{remove_tab}
\alias{reset_tab}
\title{Manage Tabbing in script creation}
\usage{
add_tab(n = 1)

remove_tab(n = 1)

reset_tab()

get_tab()
}
\arguments{
\item{n}{Number of tabs to add}

\item{n}{Number of tabs to remove}
}
\description{
Indent code for easier reading and debugging.
}
\details{
These help in code generation.

Proper use of these functions result in well indented scripts during metaprogrammiung.
}

