% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branching_logic.R
\name{toproper}
\alias{toproper}
\title{Convert strings to proper case.}
\usage{
toproper(x, all = FALSE)
}
\arguments{
\item{x}{A string.}

\item{all}{Whether to convert all words. Default is only the first word is converted to proper case, the rest is lower case.}
}
\value{
A string formatted to proper case
}
\description{
Convert strings to proper case.
}
